+++ src/main/java/com/mobble/mobbleserver/infrastructure/persistence/meetingMember/JpaMeetingMemberRepository.java
@@ -1,4 +1,4 @@
-package com.mobble.mobbleserver.refactor.meetingMember.repository;
+package com.mobble.mobbleserver.infrastructure.persistence.meetingMember;
 
 import com.mobble.mobbleserver.domain.meetingMember.MeetingMember;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface MeetingMemberRepository extends JpaRepository<MeetingMember, Long> {
+public interface JpaMeetingMemberRepository extends JpaRepository<MeetingMember, Long> {
 
     Optional<MeetingMember> findMeetingMemberByMeetingIdAndMemberId(Long meetingId, Long memberId);
