+++ src/main/java/com/mobble/mobbleserver/refactor/meetingMember/validator/MeetingMemberValidator.java
@@ -1,7 +1,7 @@
 package com.mobble.mobbleserver.refactor.meetingMember.validator;
 
 import com.mobble.mobbleserver.domain.meetingMember.MeetingMember;
-import com.mobble.mobbleserver.refactor.meetingMember.repository.MeetingMemberRepository;
+import com.mobble.mobbleserver.infrastructure.persistence.meetingMember.JpaMeetingMemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
@@ -12,7 +12,7 @@ import java.util.Optional;
 @RequiredArgsConstructor
 public class MeetingMemberValidator {
 
-    private final MeetingMemberRepository meetingMemberRepository;
+    private final JpaMeetingMemberRepository meetingMemberRepository;
 
     public Optional<MeetingMember> findMeetingByMeetingIdAndMemberId(Long meetingId, Long memberId) {
         return meetingMemberRepository.findMeetingMemberByMeetingIdAndMemberId(meetingId, memberId);
