+++ src/test/java/com/mobble/mobbleserver/refactor/meetingMember/repository/MeetingMemberRepositoryTest.java
@@ -1,6 +1,7 @@
 package com.mobble.mobbleserver.refactor.meetingMember.repository;
 
 import com.mobble.mobbleserver.config.QueryDslConfig;
+import com.mobble.mobbleserver.infrastructure.persistence.meetingMember.JpaMeetingMemberRepository;
 import com.mobble.mobbleserver.refactor.club.core.entity.Club;
 import com.mobble.mobbleserver.refactor.clubCategory.entity.ClubCategory;
 import com.mobble.mobbleserver.refactor.clubMember.entity.ClubMember;
@@ -32,7 +33,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 class MeetingMemberRepositoryTest {
 
     @Autowired
-    private MeetingMemberRepository meetingMemberRepository;
+    private JpaMeetingMemberRepository meetingMemberRepository;
 
     @Autowired
     private EntityManager em;
